+++ analysis/druidrestoration/src/CombatLogParser.ts
@@ -15,7 +15,6 @@ import Abilities from './modules/Abilities';
 import HotAttributor from './modules/core/hottracking/HotAttributor';
 import HotTrackerRestoDruid from './modules/core/hottracking/HotTrackerRestoDruid';
 import Mastery from './modules/core/Mastery';
-import Rejuvenation from './modules/core/Rejuvenation';
 import SpellManaCost from './modules/core/SpellManaCost';
 import AlwaysBeCasting from './modules/features/AlwaysBeCasting';
 import AverageHots from './modules/features/AverageHots';
@@ -27,8 +26,8 @@ import Innervate from './modules/features/Innervate';
 import Ironbark from './modules/features/Ironbark';
 import Lifebloom from './modules/features/Lifebloom';
 import LifebloomAndEffloUptime from './modules/features/LifebloomAndEffloUptime';
-import PrematureRejuvenations from './modules/features/PrematureRejuvenations';
 import RegrowthAndClearcasting from './modules/features/RegrowthAndClearcasting';
+import Rejuvenation from './modules/features/Rejuvenation';
 import RestoDruidHealingEfficiencyDetails from './modules/features/RestoDruidHealingEfficiencyDetails';
 import HealingEfficiencyTracker from './modules/features/RestoDruidHealingEfficiencyTracker';
 import StatWeights from './modules/features/StatWeights';
@@ -78,7 +77,6 @@ class CombatLogParser extends CoreCombatLogParser {
     swiftmendNormazlier: SwiftmendNormalizer,
 
     // Core
-    rejuvenation: Rejuvenation,
     mastery: Mastery,
     spellManaCost: SpellManaCost,
     activeDruidForm: ActiveDruidForm,
@@ -108,7 +106,7 @@ class CombatLogParser extends CoreCombatLogParser {
     springBlossoms: SpringBlossoms,
     cultivation: Cultivation,
     ironbark: Ironbark,
-    prematureRejuvenations: PrematureRejuvenations,
+    prematureRejuvenations: Rejuvenation,
     lifebloomAndEffloUptime: LifebloomAndEffloUptime,
     swiftmend: Swiftmend,
     hotCountGraph: HotCountGraph,
