+++ analysis/druidrestoration/src/modules/talents/TreeOfLife.tsx
@@ -23,8 +23,8 @@ import STATISTIC_ORDER from 'parser/ui/STATISTIC_ORDER';
 
 import { ABILITIES_AFFECTED_BY_HEALING_INCREASES_SPELL_OBJECTS } from '../../constants';
 import HotTrackerRestoDruid from '../../modules/core/hottracking/HotTrackerRestoDruid';
+import Rejuvenation from '../../modules/features/Rejuvenation';
 import { isFromHardcast } from '../../normalizers/CastLinkNormalizer';
-import Rejuvenation from '../core/Rejuvenation';
 
 const ALL_BOOST = 0.15;
 const ALL_MULT = 1.15;
